Move string interpolation into the compiler
This makes more sense because it allows us to use locals normally, and
lets us do less silly environment manipulation.

Note that the interpolation function is incredibly naive and may be
insufficient in some cases, but I'm adding this anyway because in any
case where interpolation breaks down, you should still be able to work
around it with regular string concatenation.

Resolves #12